improvements

+++ modules/lib/handleImages.js
@@ -7,23 +7,25 @@ const fs = require("fs");
 
 module.exports = async (item, logger) => {
   let id, palette;
+  const imgPath = `./public/images/${item.id}.jpg`;
 
   if (item.img || !item.palette) {
-    const imgPath = `./public/images/${item.id}.jpg`;
     logger(`downloading image`);
-
-    if (item.img) {
-      await download.image({
-        url: item.img.replace("/64s/", "/128s/"),
-        dest: imgPath,
-      });
+    if (!fs.existsSync(imgPath)) {
+      if (item.img) {
+        await download.image({
+          url: item.img.replace("/64s/", "/128s/"),
+          dest: imgPath,
+        });
+      }
     }
 
-    await sharp(imgPath).resize({ width: 500 }).toFile(`${imgPath}-temp`);
-    fs.unlinkSync(imgPath);
-    fs.renameSync(`${imgPath}-temp`, imgPath);
-    await webp.cwebp(imgPath, imgPath.replace(".jpg", ".webp"), "-q 75");
-
+    if (!fs.existsSync(imgPath.replace(".jpg", ".webp"))) {
+      await sharp(imgPath).resize({ width: 500 }).toFile(`${imgPath}-temp`);
+      fs.unlinkSync(imgPath);
+      fs.renameSync(`${imgPath}-temp`, imgPath);
+      await webp.cwebp(imgPath, imgPath.replace(".jpg", ".webp"), "-q 75");
+    }
     await new Promise(resolve => setTimeout(resolve, 1500));
 
     palette = await getColors(imgPath, logger);
